12-Story-Telling-Multi-Agent/requirements.txt
livekit-agents>=1.1.6
python-dotenv
ollama
faster-whisper
torch>=1.12.1
numpy
soundfile
aiohttp
sounddevice